+++ Blinkiis_Portraits/Locales/koKR.lua
@@ -132,8 +132,8 @@ L["Custom "] = "Custom "
 
 L["ERROR 1 - This is not a Blinkiis Portraits profile!"] = "ERROR 1 - This is not a Blinkii´s Portraits profile!"
 L["ERROR 2 - Import string is corrupted!"] = "ERROR 2 - Import string is corrupted!"
-L["ERROR 3 - Import string is corrupted!"] = "ERROR 3 - Import string is corrupted!"
-L["ERROR 4 - Import string is corrupted!"] = "ERROR 4 - Import string is corrupted!"
+L["ERROR - This profile was exported with an older version, please export it again."] = "ERROR - This profile was exported with an older version, please export it again."
+L["ERROR - This game version does not support profile import and export."] = "ERROR - This game version does not support profile import and export."
 
 L["BACKGROUND"] = "BACKGROUND"
 L["LOW"] = "LOW"